requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
djangorestframework==3.13.1
frozenlist==1.3.0
idna==3.3
lxml==4.8.0
multidict==6.0.2
pytz==2021.3
requests==2.27.1
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
yarl==1.7.2
